query-understanding/requirements.txt
anyio==3.6.1
asgiref==3.5.2
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.3.0/es_core_news_sm-3.3.0-py3-none-any.whl
fastapi==0.78.0
gensim==4.2.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
langdetect==1.0.9
MarkupSafe==2.1.1
murmurhash==1.0.7
nltk==3.7
numpy==1.22.3
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.9
pyspellchecker==0.6.3
pyzmq==23.0.0
regex==2022.4.24
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
snowballstemmer==2.2.0
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
starlette==0.19.1
thinc==8.0.15
threadpoolctl==3.1.0
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
wasabi==0.9.1
zmq==0.0.0
googletrans==4.0.0-rc1
